+++ core/src/main/java/eu/unicore/ucc/actions/job/Exec.java
@@ -165,7 +165,6 @@ public class Exec extends ActionBase {
 		try{
 			builder = new UCCBuilder(registry, configurationProvider);
 			builder.setProperty("Output",output.getAbsolutePath());
-			builder.setProperty("IDLocation",output.getAbsolutePath());
 			builder.setProperty("KeepFinishedJob", String.valueOf(keep));
 			builder.setProperty("DetailedStatusDisplay", "true");
 			if(tags!=null&&tags.length>0) {